# services/api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{orgId}/members": {
      "post": {
        "operationId": "MembersController_addMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Add a user to an organization",
        "tags": [
          "Organization Members"
        ]
      },
      "get": {
        "operationId": "MembersController_listMembers",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List members of an organization",
        "tags": [
          "Organization Members"
        ]
      }
    },
    "/api/v1/organizations/{orgId}/members/{userId}": {
      "delete": {
        "operationId": "MembersController_removeMember",
        "parameters": [
          {
            "name": "orgId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Remove a user from an organization",
        "tags": [
          "Organization Members"
        ]
      }
    },
...
  },
  "info": {
    "title": "Project Organizer API",
    "description": "Enterprise Event Management Platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddMemberDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
...
    }
  }
}